chore(localize): split date related functions into their own files
