+++ packages/localize/test/date/formatDate.test.js
@@ -1,14 +1,15 @@
 import { expect } from '@open-wc/testing';
-import { localize } from '../src/localize.js';
-import { localizeTearDown } from '../test-helpers.js';
+import { localize } from '../../src/localize.js';
+import { localizeTearDown } from '../../test-helpers.js';
 
-import { formatDate, parseDate, getDateFormatBasedOnLocale } from '../src/formatDate.js';
-
-beforeEach(() => {
-  localizeTearDown();
-});
+import { formatDate } from '../../src/date/formatDate.js';
+import { parseDate } from '../../src/date/parseDate.js';
 
 describe('formatDate', () => {
+  beforeEach(() => {
+    localizeTearDown();
+  });
+
   it('displays the appropriate date based on locale', async () => {
     const testDate = new Date('2012/05/21');
 
@@ -110,51 +111,5 @@ describe('formatDate', () => {
   it('returns a zero-date when input is not a Date object', async () => {
     const date = '1-1-2016';
     expect(formatDate(date)).to.equal('0000-00-00');
-  });
-});
-
-describe('parseDate()', () => {
-  function equalsDate(value, date) {
-    return (
-      Object.prototype.toString.call(value) === '[object Date]' && // is Date Object
-      value.getDate() === date.getDate() && // day
-      value.getMonth() === date.getMonth() && // month
-      value.getFullYear() === date.getFullYear() // year
-    );
-  }
-
-  afterEach(() => {
-    // makes sure that between tests the localization is reset to default state
-    document.documentElement.lang = 'en-GB';
-  });
-  it('adds leading zeros', () => {
-    expect(equalsDate(parseDate('1-1-1979'), new Date('1979/01/01'))).to.equal(true);
-    expect(equalsDate(parseDate('1-11-1979'), new Date('1979/11/01'))).to.equal(true);
-  });
-  it('creates a date object', () => {
-    expect(parseDate('10/10/2000') instanceof Date).to.equal(true);
-  });
-  it('returns a date object', () => {
-    expect(equalsDate(parseDate('1-1-1979'), new Date('1979/01/01'))).to.equal(true);
-    expect(equalsDate(parseDate('31.12.1970'), new Date('1970/12/31'))).to.equal(true);
-  });
-  it('handles all kind of delimiters', () => {
-    expect(equalsDate(parseDate('12.12.1976'), new Date('1976/12/12'))).to.equal(true);
-    expect(equalsDate(parseDate('13.12.1976'), new Date('1976/12/13'))).to.equal(true);
-    expect(equalsDate(parseDate('14.12.1976'), new Date('1976/12/14'))).to.equal(true);
-    expect(equalsDate(parseDate('14.12-1976'), new Date('1976/12/14'))).to.equal(true);
-    expect(equalsDate(parseDate('14-12/1976'), new Date('1976/12/14'))).to.equal(true);
-  });
-  it('return undefined when no valid date provided', () => {
-    expect(parseDate('12.12.1976.,')).to.equal(undefined);
-  });
-});
-
-describe('getDateFormatBasedOnLocale()', () => {
-  it('returns the positions of day, month and year', async () => {
-    localize.locale = 'en-GB';
-    expect(getDateFormatBasedOnLocale()).to.equal('day-month-year');
-    localize.locale = 'en-US';
-    expect(getDateFormatBasedOnLocale()).to.equal('month-day-year');
   });
 });
